Style main landing page image
Learn how to create transparent backgrounds for images in GIMP, apply
this knowledge to main landing page image and add it onto site

Because the white background of the image is always visible, it
only serves to ruin the appearance of the landing page. It is
inconsistent with the other elements of the page and does not match
the current theme. Get rid of white background and find solution
regarding how to fade image to transparent like a gradient

As a result, overall appearance of landing page is cleaner
and matches more closely with Letterboxd design which is used
as reference