Add content for header
Semantic markup is important to ensure code is both human and
computer readable. As such, use the correct HTML tags to make content
more accessible for machine and human visitors

Content for header uses h1 and h3 heading elements. Use h1 to indicate
that its contents constitute the most important heading of both
the heading and the landing page itself

Note that these need to be fixed so that image does not overlap them